feat: auth type 분리
